Partially solved shader issue and laggy camera
Shader issued caused by float inaccuracy with 16:9 ratio.
sorta resolved with a round-it works a lot better, but there are some artifacts still there.
Laggy camera was moved to LateUpdate, but that didn't solve all the issues with it. Slowed down the linear interpolation, sorta fixed it, but there still is some snapping going on. My computer was dying tho so it might just be a low-end thing.